enemy sway added + better code organization
sway added to enemy::update function. in this function the enemy that is selected from a list has its position updated. there's also a sway added which makes use of a sin calculation and time to create a consistent wave and make the enemy sway left to right while falling down

further more I've divided up the player, enemy and game manager code between 3 classes instead of having it all in 1 single class.
- enemy class handles enemy logic and creation
- player class handles player logic and creation
- parachute game handles all game logic such as: rendering, point count, begin/end of game loop and executing scripts of the other classes.